+++ tests/StackExchange.Redis.Tests/Naming.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
@@ -138,7 +139,14 @@ public class Naming : TestBase
             var pFrom = method.GetParameters();
             Type[] args = pFrom.Select(x => x.ParameterType).ToArray();
             Log("Checking: {0}.{1}", from.Name, method.Name);
-            Assert.Equal(typeof(CommandFlags), args.Last());
+            if (method.GetCustomAttribute<EditorBrowsableAttribute>() is EditorBrowsableAttribute attr && attr.State == EditorBrowsableState.Never)
+            {
+                // For compatibility overloads, explicitly don't ensure CommandFlags is last
+            }
+            else
+            {
+                Assert.Equal(typeof(CommandFlags), args.Last());
+            }
             var found = to.GetMethod(huntName, flags, null, method.CallingConvention, args, null);
             Assert.NotNull(found); // "Found " + name + ", no " + huntName
             var pTo = found.GetParameters();